修改查询渠道商，修改外部渠道商显示，查询成本时只查成本大于0的记录

+++ KHW/src/com/maimob/server/db/entity/OptimizationTask.java
@@ -28,7 +28,7 @@ public class OptimizationTask implements Serializable{
 		
 		this.id = Long.parseLong(task.get("id"));
 		if(task.get("channelId") != null)
-		this.channelId = Integer.parseInt(task.get("channelId"));
+		this.channelId = Long.parseLong(task.get("channelId"));
 		this.channel = task.get("channel");
 		this.startDate = task.get("startDate");
 		this.endDate = task.get("endDate");
